Move database files into data package
