+++ api/src/main/kotlin/org/pointyware/xyz/api/data/PaymentRepository.kt
@@ -1,4 +1,4 @@
-package org.pointyware.xyz.api.databases
+package org.pointyware.xyz.api.data
 
 /**
  *
