+++ api/src/main/kotlin/org/pointyware/xyz/api/data/PostgresConnectionFactory.kt
@@ -1,4 +1,4 @@
-package org.pointyware.xyz.api.databases
+package org.pointyware.xyz.api.data
 
 import org.pointyware.xyz.api.BuildConfig
 import java.sql.Connection
