+++ api/src/main/kotlin/services/UserService.kt
@@ -1,6 +1,6 @@
 package org.pointyware.xyz.api.services
 
-import org.pointyware.xyz.api.databases.AuthRepository
+import org.pointyware.xyz.api.data.AuthRepository
 import org.pointyware.xyz.core.data.dtos.Authorization
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
